// client/src/shell/paths.ts
export enum Paths {
  LOGIN = '/login',
  USER_PROFILE = '/user_profile',

  PROFILE = '/profile',
  PROFILE_DASHBOARD = '/profile/dashboard',
  PROFILE_MAP = '/profile/map',
  PROFILE_LIST = '/profile/list',

  VEHICLE = '/vehicle',
  VEHICLE_DASHBOARD = '/vehicle/dashboard',
  VEHICLE_MAP = '/vehicle/map',
  VEHICLE_LIST = '/vehicle/list',
}